jit: guard INT_MIN / -1, which LLVM sdiv leaves poison
The interpreter throws "division overflow" for INT_MIN / -1 and answers 0 for
INT_MIN % -1; the JIT emitted a raw sdiv/srem, which LLVM leaves poison. The
division now throws like the interpreter, and the modulo selects a divisor of
1, since INT_MIN % 1 is 0 - the same answer. Signed scalars only: the
interpreter does not guard vector division either, and int8/int16 have no
division at all. Repins LLVM_JIT_EMITTER_HASH.

The interpreter's DAS_FAST_INTEGER_MOD trick had the same divergence in
disguise: int32_t(A/B) on the 2^31 quotient truncates to INT_MIN on x86
and saturates to INT_MAX on arm64. Converting the quotient through
int64_t keeps every step exact on both, with no extra branch.

+++ include/daScript/simulate/sim_policy.h
@@ -142,12 +142,12 @@ namespace  das {
          static __forceinline int32_t Mod ( int32_t a, int32_t b, Context & context, LineInfo * at ) {
             if ( b==0 ) context.throw_error_at(at, "division by zero in modulo");
             double A = a, B = b;
-            return a - int32_t(A/B)*b;
+            return int32_t(a - int64_t(A/B)*b);
         }
         static __forceinline void SetMod    ( int32_t & a, int32_t b, Context & context, LineInfo * at ) {
             if ( b==0 ) context.throw_error_at(at, "division by zero in modulo");
             double A = a, B = b;
-            a = a - int32_t(A/B)*b;
+            a = int32_t(a - int64_t(A/B)*b);
         }
     };
